api/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
kagglehub
tqdm
python-multipart
fastapi
uvicorn
python-multipart
torch
torchvision
timm
pillow
kagglehub
